refactor : AnswerRepository 의존성 역전 적용
